Fix isAllowedValue
