+++ packages/dds/tree/src/test/feature-libraries/valueUtilities.spec.ts
@@ -11,6 +11,7 @@ import { ValueSchema } from "../../core/index.js";
 import {
 	allowsValue,
 	isFluidHandle,
+	isTreeValue,
 	// Allow importing from this specific file which is being tested:
 	/* eslint-disable-next-line import/no-internal-modules */
 } from "../../feature-libraries/valueUtilities.js";
@@ -76,5 +77,19 @@ describe("valueUtilities", () => {
 		assert(!allowsValue(ValueSchema.String, null));
 		assert(!allowsValue(ValueSchema.Number, null));
 		assert(allowsValue(ValueSchema.Null, null));
+	});
+
+	it("isTreeValue", () => {
+		assert(isTreeValue(0));
+		assert(isTreeValue(0.001));
+		assert(isTreeValue(NaN));
+		assert(isTreeValue(true));
+		assert(isTreeValue(false));
+		assert(isTreeValue(""));
+		assert(!isTreeValue({}));
+		assert(!isTreeValue(undefined));
+		assert(isTreeValue(null));
+		assert(!isTreeValue([]));
+		assert(isTreeValue(new MockHandle(5)));
 	});
 });
